+++ website/src/components/Credits.tsx
@@ -92,8 +92,9 @@ export const Credits = component$(() => {
           'fabulousgk',
           'jdgamble555',
           'isoden',
+          'nickytonline',
           'caegdeveloper',
-          'andrew-d-jackson',
+          'andrew-3kb',
           'dslatkin',
         ].map((sponsor) => (
           <li key={sponsor}>