+++ library/CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to the library will be documented in this file.
 - Change type signature of `partialCheck` and `partialCheckAsync` action to add `.pathList` property in a type-safe way
 - Change type signature of `findItem` action to support type predicates (issue #867)
 - Refactor `bytes`, `maxBytes`, `minBytes` and `notBytes` action
+- Fix implementation of `nonOptional`, `nonOptionalAsync`, `nonNullable`, `nonNullableAsync`, `nonNullish` and `nonNullishAsync` schema in edge cases (issue #909)
 
 ## v0.42.1 (September 20, 2024)